Do not pass entity id to XTDB when current entity is missing
This patch fixes a bug that was causing XTDB to throw when we could not find the current entity in
the db and `:atomic?` is true (aka we use an ::xt/match).

We now take the :xt/id from *either* the current and the new entity.

This patch also removes the unnecessary `(filterv some?)` - XTDB already handles `nil`s gracefully.